Update & Delete a User Goal

+++ backend/controllers/goalController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 // @desc  Get goals
 // @route GET /api/goals
@@ -33,6 +34,18 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error('Goal not found');
   }
+  const user = await User.findById(req.user.id);
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -54,6 +67,17 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if (!goal) {
       res.status(404);
       throw new Error('Goal not found');
+    }
+    const user = await User.findById(req.user.id);
+    // Check for user
+    if (!user) {
+      res.status(401);
+      throw new Error('User not found');
+    }
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+      res.status(401);
+      throw new Error('User not authorized');
     }
 
     await goal.deleteOne();
